Move test classes to appropriate package
Leftover from #549.
